+++ backend/Mizan.Application/Queries/GetRecipesQuery.cs
@@ -40,7 +40,7 @@ public record RecipeDto
 
 public record RecipeNutritionDto
 {
-    public int? CaloriesPerServing { get; init; }
+    public decimal? CaloriesPerServing { get; init; }
     public decimal? ProteinGrams { get; init; }
     public decimal? CarbsGrams { get; init; }
     public decimal? FatGrams { get; init; }
